Move dcs implementations into dcs package
