+++ patroni/dcs/etcd.py
